clean up agent node attestation tests
